Moved files from web_site to people app
