+++ acm_web_site/apps/business/people/serializer/member_serializer.py
@@ -5,7 +5,7 @@ import base64
 from django.conf import settings
 
 from rest_framework import serializers
-from ..models import Member, POSITION_CHOICES
+from acm_web_site.apps.business.web_site.models import Member, POSITION_CHOICES
 
 
 class MemberSerializer(serializers.ModelSerializer):
