+++ acm_web_site/apps/business/people/urls.py
@@ -1,0 +1,12 @@
+from django.urls import path
+from . import views
+
+
+handler404 = 'apps.business.web_site.views.page_not_found'
+
+urlpatterns = [
+	path('members', views.members),
+	path('teams/', views.TeamList.as_view()),
+	path('joinus/', views.join_us),
+
+]
