+++ acm_web_site/urls.py
@@ -6,5 +6,6 @@ from markdownx import urls as markdownx
 urlpatterns = [
 	path('admin/', admin.site.urls),
 	path('', include('apps.business.web_site.urls')),
+	path('', include('apps.business.people.urls')),
 	path('markdownx/', include('markdownx.urls')),
 ]
